jobs: deprecate node liveness functions and data
Release note: none.
