+++ pkg/jobs/registry.go
@@ -14,7 +14,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"math/rand"
 	"os"
 	"strings"
 	"time"
@@ -24,7 +23,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/settings"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
@@ -55,13 +53,6 @@ const defaultLeniencySetting = 60 * time.Second
 const multiTenancyIssueNo = 47892
 
 var (
-	nodeLivenessLogLimiter = log.Every(5 * time.Second)
-	// LeniencySetting is the amount of time to defer any attempts to
-	// reschedule a job.  Visible for testing.
-	LeniencySetting = settings.RegisterDurationSetting(
-		"jobs.registry.leniency",
-		"the amount of time to defer any attempts to reschedule a job",
-		defaultLeniencySetting)
 	gcSetting = settings.RegisterPublicDurationSetting(
 		"jobs.retention_time",
 		"the amount of time to retain records for completed jobs before",
@@ -144,7 +135,7 @@ type Registry struct {
 		// epoch is present to support older nodes that are not using
 		// the timestamp-based approach to determine when to steal jobs.
 		// TODO: Remove this and deprecate Lease.Epoch proto field
-		epoch int64
+		deprecatedEpoch int64
 		// jobs holds a map from job id to its context cancel func. This should
 		// be populated with jobs that are currently being run (and owned) by
 		// this registry. Calling the func will cancel the context the job was
@@ -154,7 +145,7 @@ type Registry struct {
 		// propagated to jobs via the .Progressed call. This function should not be
 		// used to cancel a job in that way.
 		// TODO(spaskob): add deprecated notice.
-		jobs map[int64]context.CancelFunc
+		deprecatedJobs map[int64]context.CancelFunc
 
 		// adoptedJobs holds a map from job id to its context cancel func and epoch.
 		// It contains the that are adopted and rpobably being run. One exception is
@@ -215,8 +206,8 @@ func MakeRegistry(
 		preventAdoptionFile: preventAdoptionFile,
 		adoptionCh:          make(chan struct{}),
 	}
-	r.mu.epoch = 1
-	r.mu.jobs = make(map[int64]context.CancelFunc)
+	r.mu.deprecatedEpoch = 1
+	r.mu.deprecatedJobs = make(map[int64]context.CancelFunc)
 	r.mu.adoptedJobs = make(map[int64]*adoptedJob)
 	r.metrics.InitHooks(histogramWindowInterval)
 	return r
@@ -261,21 +252,6 @@ func (r *Registry) CurrentlyRunningJobs() []int64 {
 // used for keying sqlliveness claims held by the registry.
 func (r *Registry) ID() base.SQLInstanceID {
 	return r.nodeID.SQLInstanceID()
-}
-
-// lenientNow returns the timestamp after which we should attempt
-// to steal a job from a node whose liveness is failing.  This allows
-// jobs coordinated by a node which is temporarily saturated to continue.
-func (r *Registry) lenientNow() time.Time {
-	// We see this in tests.
-	var offset time.Duration
-	if r.settings == cluster.NoSettings {
-		offset = defaultLeniencySetting
-	} else {
-		offset = LeniencySetting.Get(&r.settings.SV)
-	}
-
-	return r.clock.Now().GoTime().Add(-offset)
 }
 
 // makeCtx returns a new context from r's ambient context and an associated
@@ -410,7 +386,7 @@ func (r *Registry) CreateJobWithTxn(ctx context.Context, record Record, txn *kv.
 		ctx, clusterversion.VersionAlterSystemJobsAddSqllivenessColumnsAddNewSystemSqllivenessTable) {
 		// TODO(spaskob): remove in 20.2 as this code path is only needed while
 		// migrating to 20.2 cluster.
-		if err := j.WithTxn(txn).insert(ctx, r.makeJobID(), r.newLease()); err != nil {
+		if err := j.WithTxn(txn).insert(ctx, r.makeJobID(), r.deprecatedNewLease()); err != nil {
 			return nil, err
 		}
 		return j, nil
@@ -517,7 +493,7 @@ func (r *Registry) CreateStartableJobWithTxn(
 	} else {
 		// TODO(spaskob): remove in 20.2 as this code path is only needed while
 		// migrating to 20.2 cluster.
-		if err := r.register(*j.ID(), cancel); err != nil {
+		if err := r.deprecatedRegister(*j.ID(), cancel); err != nil {
 			return nil, err
 		}
 	}
@@ -611,10 +587,10 @@ func (r *Registry) Start(
 
 	maybeAdoptJobs := func(ctx context.Context, randomizeJobOrder bool) {
 		if r.adoptionDisabled(ctx) {
-			r.cancelAll(ctx)
+			r.deprecatedCancelAll(ctx)
 			return
 		}
-		if err := r.maybeAdoptJob(ctx, r.nl, randomizeJobOrder); err != nil {
+		if err := r.deprecatedMaybeAdoptJob(ctx, r.nl, randomizeJobOrder); err != nil {
 			log.Errorf(ctx, "error while adopting jobs: %s", err)
 		}
 	}
@@ -690,7 +666,7 @@ func (r *Registry) maybeCancelJobs(ctx context.Context, nlw sqlbase.OptionalNode
 	// Cancel all jobs if the stopper is quiescing.
 	select {
 	case <-r.stopper.ShouldQuiesce():
-		r.cancelAll(ctx)
+		r.deprecatedCancelAll(ctx)
 		log.Warningf(ctx, "canceling all adopted jobs due to stopper quiescing")
 		r.cancelAllAdoptedJobs()
 		return
@@ -703,7 +679,7 @@ func (r *Registry) maybeCancelJobs(ctx context.Context, nlw sqlbase.OptionalNode
 		defer r.mu.Unlock()
 		// If the cluster is finalized, kill any remaining legacy jobs. They will be
 		// re-adopted with the new epoch based leasing.
-		r.cancelAllLocked(ctx)
+		r.deprecatedCancelAllLocked(ctx)
 
 		// This gets the current live epoch and pushes forward its expiration. If
 		// another regitry has bumped it, we don;t need to know - as it suffices to
@@ -740,7 +716,7 @@ func (r *Registry) maybeCancelJobs(ctx context.Context, nlw sqlbase.OptionalNode
 				log.Warningf(ctx, "unable to get node liveness: %s", err)
 			}
 			// Conservatively assume our lease has expired. Abort all jobs.
-			r.cancelAll(ctx)
+			r.deprecatedCancelAll(ctx)
 			return
 		}
 
@@ -749,8 +725,8 @@ func (r *Registry) maybeCancelJobs(ctx context.Context, nlw sqlbase.OptionalNode
 		if !liveness.IsLive(r.lenientNow()) {
 			r.mu.Lock()
 			defer r.mu.Unlock()
-			r.cancelAllLocked(ctx)
-			r.mu.epoch = liveness.Epoch
+			r.deprecatedCancelAllLocked(ctx)
+			r.mu.deprecatedEpoch = liveness.Epoch
 			return
 		}
 	}
@@ -1132,7 +1108,6 @@ func (r *Registry) resume(
 	ctx context.Context, resumer Resumer, resultsCh chan<- tree.Datums, job *Job, onDone func(),
 ) (<-chan error, error) {
 	errCh := make(chan error, 1)
-	taskName := fmt.Sprintf(`job-%d`, *job.ID())
 	if err := r.stopper.RunAsyncTask(ctx, job.taskName(), func(ctx context.Context) {
 		if onDone != nil {
 			defer onDone()
@@ -1198,255 +1173,6 @@ func (r *Registry) adoptionDisabled(ctx context.Context) bool {
 	return false
 }
 
-func (r *Registry) maybeAdoptJob(
-	ctx context.Context, nlw sqlbase.OptionalNodeLiveness, randomizeJobOrder bool,
-) error {
-	const stmt = `
-SELECT id, payload, progress IS NULL, status
-FROM system.jobs
-WHERE status IN ($1, $2, $3, $4, $5) ORDER BY created DESC`
-	rows, err := r.ex.Query(
-		ctx, "adopt-job", nil /* txn */, stmt,
-		StatusPending, StatusRunning, StatusCancelRequested, StatusPauseRequested, StatusReverting,
-	)
-	if err != nil {
-		return errors.Wrap(err, "failed querying for jobs")
-	}
-
-	if randomizeJobOrder {
-		rand.Seed(timeutil.Now().UnixNano())
-		rand.Shuffle(len(rows), func(i, j int) { rows[i], rows[j] = rows[j], rows[i] })
-	}
-
-	type nodeStatus struct {
-		isLive bool
-	}
-	nodeStatusMap := map[roachpb.NodeID]*nodeStatus{
-		// We treat invalidNodeID as an always-dead node so that
-		// the empty lease (Lease{}) is always considered expired.
-		invalidNodeID: {isLive: false},
-	}
-	// If no liveness is available, adopt all jobs. This is reasonable because this
-	// only affects SQL tenants, which have at most one SQL server running on their
-	// behalf at any given time.
-	if nl, ok := nlw.Optional(47892); ok {
-		// We subtract the leniency interval here to artificially
-		// widen the range of times over which the job registry will
-		// consider the node to be alive.  We rely on the fact that
-		// only a live node updates its own expiration.  Thus, the
-		// expiration time can be used as a reasonable measure of
-		// when the node was last seen.
-		now := r.lenientNow()
-		for _, liveness := range nl.GetLivenesses() {
-			nodeStatusMap[liveness.NodeID] = &nodeStatus{
-				isLive: liveness.IsLive(now),
-			}
-
-			// Don't try to start any more jobs unless we're really live,
-			// otherwise we'd just immediately cancel them.
-			if liveness.NodeID == r.nodeID.DeprecatedNodeID(multiTenancyIssueNo) {
-				if !liveness.IsLive(r.clock.Now().GoTime()) {
-					return errors.Errorf(
-						"trying to adopt jobs on node %d which is not live", liveness.NodeID)
-				}
-			}
-		}
-	}
-
-	if log.V(3) {
-		log.Infof(ctx, "evaluating %d jobs for adoption", len(rows))
-	}
-
-	var adopted int
-	for _, row := range rows {
-		if adopted >= maxAdoptionsPerLoop {
-			// Leave excess jobs for other nodes to get their fair share.
-			break
-		}
-
-		id := (*int64)(row[0].(*tree.DInt))
-
-		payload, err := UnmarshalPayload(row[1])
-		if err != nil {
-			return err
-		}
-
-		status := Status(tree.MustBeDString(row[3]))
-		if log.V(3) {
-			log.Infof(ctx, "job %d: evaluating for adoption with status `%s` and lease %v",
-				*id, status, payload.Lease)
-		}
-
-		// In version 20.1, the registry must not adopt 19.2-style schema change
-		// jobs until they've undergone a migration.
-		// TODO (lucy): Remove this in 20.2.
-		if isOldSchemaChangeJob(payload) {
-			log.VEventf(ctx, 2, "job %d: skipping adoption because schema change job has not been migrated", id)
-			continue
-		}
-
-		if payload.Lease == nil {
-			// If the lease is missing, it simply means the job does not yet support
-			// resumability.
-			if log.V(2) {
-				log.Infof(ctx, "job %d: skipping: nil lease", *id)
-			}
-			continue
-		}
-
-		// If the job has no progress it is from a 2.0 cluster. If the entire cluster
-		// has been upgraded to 2.1 then we know nothing is running the job and it
-		// can be safely failed.
-		if nullProgress, ok := row[2].(*tree.DBool); ok && bool(*nullProgress) {
-			log.Warningf(ctx, "job %d predates cluster upgrade and must be re-run", *id)
-			versionErr := errors.New("job predates cluster upgrade and must be re-run")
-			payload.Error = versionErr.Error()
-			payloadBytes, err := protoutil.Marshal(payload)
-			if err != nil {
-				return err
-			}
-
-			// We can't use job.update here because it fails while attempting to unmarshal
-			// the progress. Setting the status to failed is idempotent so we don't care
-			// if multiple nodes execute this.
-			const updateStmt = `UPDATE system.jobs SET status = $1, payload = $2 WHERE id = $3`
-			updateArgs := []interface{}{StatusFailed, payloadBytes, *id}
-			err = r.db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) error {
-				_, err := r.ex.Exec(ctx, "job-update", txn, updateStmt, updateArgs...)
-				return err
-			})
-			if err != nil {
-				log.Warningf(ctx, "job %d: has no progress but unable to mark failed: %s", *id, err)
-			}
-			continue
-		}
-
-		r.mu.Lock()
-		_, runningOnNode := r.mu.jobs[*id]
-		r.mu.Unlock()
-
-		// If we're running as a tenant (!ok), then we are the sole SQL server in
-		// charge of its jobs and ought to adopt all of them. Otherwise, look more
-		// closely at who is running the job and whether to adopt.
-		if nodeID, ok := r.nodeID.OptionalNodeID(); ok && nodeID != payload.Lease.NodeID {
-			// Another node holds the lease on the job, see if we should steal it.
-			if runningOnNode {
-				// If we are currently running a job that another node has the lease on,
-				// stop running it.
-				log.Warningf(ctx, "job %d: node %d owns lease; canceling", *id, payload.Lease.NodeID)
-				r.unregister(*id)
-				continue
-			}
-			nodeStatus, ok := nodeStatusMap[payload.Lease.NodeID]
-			if !ok {
-				// This case should never happen.
-				log.ReportOrPanic(ctx, nil, "job %d: skipping: no liveness record for the job's node %d",
-					log.Safe(*id), payload.Lease.NodeID)
-				continue
-			}
-			if nodeStatus.isLive {
-				if log.V(2) {
-					log.Infof(ctx, "job %d: skipping: another node is live and holds the lease", *id)
-				}
-				continue
-			}
-		}
-
-		// Below we know that this node holds the lease on the job, or that we want
-		// to adopt it anyway because the leaseholder seems dead.
-		job := &Job{id: id, registry: r}
-		resumeCtx, cancel := r.makeCtx()
-
-		if pauseRequested := status == StatusPauseRequested; pauseRequested {
-			if err := job.paused(ctx, func(context.Context, *kv.Txn) error {
-				r.unregister(*id)
-				return nil
-			}); err != nil {
-				log.Errorf(ctx, "job %d: could not set to paused: %v", *id, err)
-				continue
-			}
-			log.Infof(ctx, "job %d: paused", *id)
-			continue
-		}
-
-		if cancelRequested := status == StatusCancelRequested; cancelRequested {
-			if err := job.reverted(ctx, errJobCanceled, func(context.Context, *kv.Txn) error {
-				// Unregister the job in case it is running on the node.
-				// Unregister is a no-op for jobs that are not running.
-				r.unregister(*id)
-				return nil
-			}); err != nil {
-				log.Errorf(ctx, "job %d: could not set to reverting: %v", *id, err)
-				continue
-			}
-			log.Infof(ctx, "job %d: canceled: the job is now reverting", *id)
-		} else if currentlyRunning := r.register(*id, cancel) != nil; currentlyRunning {
-			if log.V(3) {
-				log.Infof(ctx, "job %d: skipping: the job is already running/reverting on this node", *id)
-			}
-			continue
-		}
-
-		// Check if job status has changed in the meanwhile.
-		currentStatus, err := job.CurrentStatus(ctx)
-		if err != nil {
-			return err
-		}
-		if status != currentStatus {
-			continue
-		}
-		// Adopt job and resume/revert it.
-		if err := job.adopt(ctx, payload.Lease); err != nil {
-			r.unregister(*id)
-			return errors.Wrap(err, "unable to acquire lease")
-		}
-
-		resultsCh := make(chan tree.Datums)
-		resumer, err := r.createResumer(job, r.settings)
-		if err != nil {
-			r.unregister(*id)
-			return err
-		}
-		log.Infof(ctx, "job %d: resuming execution", *id)
-		errCh, err := r.resume(resumeCtx, resumer, resultsCh, job, nil)
-		if err != nil {
-			r.unregister(*id)
-			return err
-		}
-		go func() {
-			// Drain and ignore results.
-			for range resultsCh {
-			}
-		}()
-		go func() {
-			// Wait for the job to finish. No need to print the error because if there
-			// was one it's been set in the job status already.
-			<-errCh
-			close(resultsCh)
-		}()
-
-		adopted++
-	}
-
-	return nil
-}
-
-func (r *Registry) newLease() *jobspb.Lease {
-	nodeID := r.nodeID.DeprecatedNodeID(multiTenancyIssueNo)
-	if nodeID == 0 {
-		panic("jobs.Registry has empty node ID")
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return &jobspb.Lease{NodeID: nodeID, Epoch: r.mu.epoch}
-}
-
-func (r *Registry) cancelAll(ctx context.Context) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.cancelAllLocked(ctx)
-}
-
 func (r *Registry) cancelAllAdoptedJobs() {
 	r.mu.Lock()
 	defer r.mu.Unlock()
@@ -1456,44 +1182,21 @@ func (r *Registry) cancelAllAdoptedJobs() {
 	r.mu.adoptedJobs = make(map[int64]*adoptedJob)
 }
 
-func (r *Registry) cancelAllLocked(ctx context.Context) {
-	r.mu.AssertHeld()
-	for jobID, cancel := range r.mu.jobs {
-		log.Warningf(ctx, "job %d: canceling due to liveness failure", jobID)
-		cancel()
-	}
-	r.mu.jobs = make(map[int64]context.CancelFunc)
-}
-
-// register registers an about to be resumed job in memory so that it can be
-// killed and that no one else tries to resume it. This essentially works as a
-// barrier that only one function can cross and try to resume the job.
-func (r *Registry) register(jobID int64, cancel func()) error {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	// We need to prevent different routines trying to adopt and resume the job.
-	if _, alreadyRegistered := r.mu.jobs[jobID]; alreadyRegistered {
-		return errors.Errorf("job %d: already registered", jobID)
-	}
-	r.mu.jobs[jobID] = cancel
-	return nil
-}
-
 func (r *Registry) unregister(jobID int64) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
-	cancel, ok := r.mu.jobs[jobID]
+	cancel, ok := r.mu.deprecatedJobs[jobID]
 	// It is possible for a job to be double unregistered. unregister is always
-	// called at the end of resume. But it can also be called during cancelAll
+	// called at the end of resume. But it can also be called during deprecatedCancelAll
 	// and in the adopt loop under certain circumstances.
 	if ok {
 		cancel()
-		delete(r.mu.jobs, jobID)
+		delete(r.mu.deprecatedJobs, jobID)
 		return
 	}
 	aj, ok := r.mu.adoptedJobs[jobID]
 	// It is possible for a job to be double unregistered. unregister is always
-	// called at the end of resume. But it can also be called during cancelAll
+	// called at the end of resume. But it can also be called during deprecatedCancelAll
 	// and in the adopt loop under certain circumstances.
 	if ok {
 		aj.cancel()
